Update patch set 23
Patch Set 23: Code-Review-1

There are a bunch of errors in RTL (and LTR) with this fix still:

* Seems to ignore slugs (both ltr/rtl)
(Testing document has block images at the top and text under the images in wikitext. The images are rendered on the side)

So, in RTL, if I write text at the top of the document (transforming a slug into paragraph) and then move the cursor from the beginning of the RTL sentence to the end, it stops at the end of that sentence and refuses to continue unless I click 'down' arrow to move forward.

In LTR, the cursor jumps from one line of text directly to the next one, ignoring slugs and ignoring the images that are supposedly the "next" item to select.

* Inline images

In ve-mw LTR this seems to work well, but in the Standalone in both LTR and RTL if you have a sentence and an inline image, the cursor goes a bit crazy after trying to "skip" over the image. It seems to select the image + the entire line and then vanishes (at least it doesn't break anymore!)

In RTL this actually works well when going from beginning of sentence to the end of the sentence, but it breaks when going from end of sentence to the beginning (flipped problem from LTR)

yrroS, David!

Patch-set: 23
Label: Code-Review=-1