Update patch set 1
Patch Set 1:

Yeah .ve-init-target-ieNN was what I was thinking about too.

I'll put this on the backburner for now, because at this point the situation we're talking about is theoretical. The only place in which it sort of applies right now is that if we wanted to, we could reapply the blue background for language annotations in IE12.

Patch-set: 1