Update patch set 25
Patch Set 25: Code-Review-1

Okay, everything seems to work browser-like now (moving the arrows is visual, marking is logical) except for one bug -- 

If you have an inline image in the middle of a sentence, the cursor gets "stuck" when you try to move through it (either from beginning to end or from end to beginning) 

This happens in both RTL and LTR, but only in Chrome. Firefox seems to handle this nicely.

In Firefox, Shift-cursoring (marking) seems to work okay for me in both LTR and RTL, and it marks images fine.

However, in Chrome, shift-cursoring works okay on text, but completely fails on inline images; if you have a sentence with an inline image, and you shift-cursor (mark) before the image, then after the selection passes through the image, the image and the selection vanish and a transaction is fired (save button is enabled if it wasn't before)

Patch-set: 25
Label: Code-Review=-1