Update patch set 27
Patch Set 27: Code-Review-1

Okay, sadly, I found a bug in both LTR/RTL, and in both Chrome and Firefox:

Inline template or inline images: Start with a sentence that has some inline template (or image) in the middle of it. Go from the beginning of the sentence and hit 'right' to proceed forward. The cursor gets "stuck" when it reaches the inline template - it alternates between marking the template and moving *back* to be in front of it. The only way to "get out" of that loop is to hit "right" to mark the template and then, counter-intuitively, hit *left* (as if you want to go backwards) -- that actually places the cursor *after* the template, and you can then continue clicking "right" to move forward.

NOTES: 
* The above is reproducible in both LTR and RTL. For RTL, just switch 'left' for 'right' and vice versa.
* *marking* text (shift-arrow) in the above case works perfectly in both Firefox and Chrome, yay!
* In Firefox/RTL, the inline template is being skipped when you move across it; so, on one hand, this bug doesn't exist there, but on the other, you can't "land" on an inline template node moving arrows. The bug still happens with inline images, however.

Patch-set: 27
Reviewer: Mooeypoo <798@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1