Update patch set 1
Patch Set 1:

I realize this isn't exactly well-documented, but the targetDoc parameter to getModelFromDom() is NOT meant to be the view document. In no case should any DM code even be aware of the view document. Instead, pass null for that parameter (which is essentially what is happening already, that's why the current code works).

Patch-set: 1