Add lookup-by-ID feature to InternalList
When queueing up an HTML entry, an optional ID can be passed.
If no non-empty HTML is ever queued up for a given key,
we will fall back to the contents of the DOM element with
this ID, if one exists. This lookup is done when converting
to linmod data, in the document passed to convertToData().

Change-Id: I2a941dfbc27374eb4029fa012a30a93e4975f1f4