Update patch set 6
Patch Set 6: Code-Review-1

I think this isn't quite where it needs to be to supplant the current, albeit broken, implementation. Keyboard accessibility is still an issue, for one, and the design treatment could probably be better to indicate the area of line being added rather than just the anchor point.

Patch-set: 6
Label: Code-Review=-1