Update patch set 1
Patch Set 1:

Yeah, but there is one case where they are appropriate: When your DOM order does not reflect the desired stacking order. If our toolbar would always be floating, then it would be appropriate for it to be last in the dom (instead of first) and thus not need z-index.

However since we want it first in the dom (so that when its not floating, its in between the tabs and surface where it should be and can be transitioned in naturally in-place), we'd need z-index to make this work.

But.. I'm implementing I30a7b69b77b874d2 another way because of limitations in the transform/translate mechanism in the browser. So this is low-priority now and not blocking. Though it is harmless and should be merged for semantic reasons. If not, it will bite us again at some point. it supposed to stack higher, even if we're not exploiting it.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0