ce.ContentBranchNode: Don't compare new rendering with old on first render
We unconditionally compared the new rendering with the old rendering,
even on the first render. This wasn't actually a no-op because, while
text nodes aren't yet in this.$element before the first render,
inline nodes and slugs are.

As an optimization, only perform this comparison if the node has
rendered itself before.

Change-Id: I7e1315e63e9022891c82f1d2e447c37974d70bd3