Don't call .addClass() in CE nodes' setup handlers
Adding CSS classes on setup causes nodes to first render without
the class, then rerender with the class. This is wasteful. Instead,
add classes in the constructor.

In LeafNode, BranchNode and FocusableNode we also set classes
on setup, because GeneratedContentNode does crazy things to this.$element
and it's important we don't lose these classes (a lot of code
checks for the presence of these classes).

In TableCellNode, we need to update classes when the style
attribute changes.

Change-Id: I2690751d16977b106831c94a0757b615985fa5d6