Make ce.BranchNode#updateTagName do what its documentation says
Actually transfer classes, and transfer the contentEditable property too.
This allows us to set classes and cE in the constructor instead
of on setup.

Also rewrite updateTagName to use DOM methods, and update TableCellNode
to have its teardown listener remove the classes added by its setup
listener (it previously got away with not doing that because updateTagName
didn't behave as documented).

Change-Id: I7547a2228754332b907f36d9705b9987b12d1091