Queue linear model modifications in TransactionProcessor
In eaa4528deb annotation changes were queued; this commit
extends that approach to the entirety of TransactionProcessor.
All linear model modifications are queued up, then executed
at the end. Modifications are expressed as method calls
on this.document.data or this.document.metadata.

This change also changes how this.cursor operates:
because the linear model now doesn't change as we step
through it, this.cursor now always corresponds to offsets
in the old, unmodified linear model, so we can now use
it for tree lookups directly without subtracting this.adjustment.
We never need to subtract this.adjustment any more, and we
only need to add it in cases where we're scheduling a deferred
action: when queueing a modification and when queueing things
for the synchronizer.

This will make it easier for us to recover cleanly
from exceptions thrown while processing transactions.

Change-Id: I09c8be26ee3dfbd6bf649d1c94862331887b61a7