Render HTML attributes in ce.View#initialize rather than in the constructor
This means that when this.$element is changed, miscellaneous attributes
aren't removed any more. For instance, if a table cell has
style="background: red;" and the user changes it to a heading cell,
the red background will no longer disappear.

(It already didn't disappear in the final version that ends up
being saved, but until now the rendering lied and didn't show the
red background.)

Change-Id: Ib552bd1e124f1bc932454fda07056d4979925b9e