Update patch set 2
Patch Set 2:

Yeah, I figured it couldn't be bound right now. But as a maintenance principle, an unbounded check should have information near it for why it is and/or a clue as to when it could be bound. Referencing our task, and (if appropiate) an upstream bug would suffice.

Patch-set: 2