Add directionality classes to window manager
Many of the icon settings require the directionality class in
order to feature icons that can be flipped between ltr/rtl. Those
icons don't show inside dialogs and inspectors unless the wrapping
window manager holds those classes.

In this fix we set the directionality based on the given surface;
however, window manager can be created (and is created) without
being given a surface; in those cases, we have to have fallbacks
to the elements on the page.

Change-Id: Ibf29417a2eb64d966417cfa96a0d4417e3d80efd