Improve backspacing over the end of a list with empty last item
It is actually possible to get rid of a list item this way, previously
all that happened was that the paragraph inside the list was being
removed and replaced by a block slug, that could be turned back into
a paragraph.

It is still possible to get a slug to appear in this position (by
creating a nested list and removing some list items), but I see
no sane solution to that.

Bug: T91349
Change-Id: I5cf49de7bf4db7ea59e8c51e6f9dcb2423352712