Expand annotation beyond a word if it exists
For the annotation inspector, if we start before a word, the
selection expands to fit the word. However, we may also be inside
a bigger annotation already. In that case, we want to run another
test to find the bigger annotation and expand the selection to
cover it as well.

Bug: T86399
Change-Id: I94a468f42e087eb16c68ba80fe74164e7a4875ee