Update patch set 3
Patch Set 3:

Citoid inspector.

When the user looks up a url, we're creating the reference node and displaying it for them. Since Citoid technically could come back with more than one option, we wait until all nodes are fully rendered.

However, as we've learned yesterday in a bout of debugging, if the user uses a node that is already create (either because they enter a url that already exists as a reference or because they clicked lookup and then cancel and then look up again) - then the node is taken from the store, and is not rerendered. The rerender event is emitted before we even have a chance to connect to that node, and the promise that the inspector is waiting on for the rendering is never resolved or rejected.

This is used to first check if the node is already generated. If it is, we skip the event, give the inspector the ready-made node, and resolve the promise. Otherwise, we create the node again and resolve the promise once the node emits 'rerender'.

Patch-set: 3