Associate the fragment in AnnotationInspector
We are manipulating the selection and fragment in the getSetupProcess
and we need to store and replace the original fragment with the new
one. This will also mean that on teardown we can refer to the actual
selection rather than an old one.

Also make sure that whenever we return/leave the teardown process
we are popping the staging.

Bug: T93363
Change-Id: Ic5802b2a249f48686387aa05a6ead6d8ac0965bb