Update patch set 4
Patch Set 4:

I don't think that's important. The only reason we highlight language annotations is to make then easier to edit. A preview mode is not for editing but for reading.
In general it depends if you think of preview as a preview of what the reference will look like, or a preview of what the reference editor will look like. I think the former makes more sense.

Patch-set: 4
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0