Simplify getClipboardHash
In WebKit browsers pasting can re-order and add spans, so getClipboardHash
was suggesting the paste data had changed when it hadn't. This is not
a problem in Chrome and Firefox where we have the clipboard API path,
but Safari doesn't support this.

Additionally, context markers also broke clipboard hash comparison in
browsers where the clipboard API path isn't used, even if spans weren't
being reordered or added. So detection of internal paste was completely
broken in IE as well as in Safari.

Simplify getClipboardHash to just return plain text without whitespace.

Bug: T71494
Change-Id: Ice5199e4db996b577d60e7df9ccb730300c9ceee