Work around IE link paste bug
IE tries to restore the appearance of links when they are being pasted
in a place where different CSS applies. If text-decoration: none; applies
to the paste target, link content will be wrapped in <u> tags, and
if a different link color applies, link content will be wrapped in
<font color="#0066cc"> tags. To prevent this from happening, explicitly
apply the default text-decoration and color styles to the paste target
so IE will refrain from mangling links this way.

Change-Id: I2b07aa934d804735316e9e736f5b25430786b1e2