Fix JS error when pasting only aliens
If you paste a block alien, the pasteTarget HTML might end up
looking something like this:

<blockquote><div rel="ve:Alien"><p>Foo</p><div><br></div></div></blockquote>

in which case the sanitization step will leave nothing behind
but the internal list, so internalListRange.start will be 0.
The code doesn't expect this, and unconditionally subtracts 1
from this number, resulting in JS errors.

Before subtracting 1, check that the value we're dealing with
really is >0.

Change-Id: Ic7884b638c692bbea9cc697dda87bfa976cb106a