Update patch set 1
Patch Set 1:

This is for the exception that Mark Hershberger found in T92969. This doesn't fix the exception itself, it just makes an exception with a sensible message be thrown instead of crashing in getClonedElement() before it can get there.

Patch-set: 1
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0