Update patch set 5
Patch Set 5: Code-Review-1

This appears to work in the editor, but it breaks in the tests. It looks like setupSlugs() isn't getting run at the right time. I don't know how to solve this, especially now that there is also a setupSlugs in the model that also needs to be called.

Patch-set: 5
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1