Trim leading and trailing whitespace from annotations
This means that even if you do try to create "<a>  Foo  </a>bar" in
the editor (and you have to try pretty hard to do that in the first place),
it'll serialize to "  <a>Foo</a>  bar".

For VE-MW, this prevents ugly wikitext output like [[Foo|Foo  ]]<nowiki/>bar.

Bug: T54037
Change-Id: I0b734651a5b39e16142cb0edc28a594ca3d60231