Trim leading and trailing whitespace from annotations
This means that even if you do try to create "<a>  Foo  </a>bar" in
the editor (and you have to try pretty hard to do that in the first place),
it'll serialize to "  <a>Foo</a>  bar".

Whitespace isn't stripped when the annotation was present in the original
document and already had whitespace there.

This is a total hack, but it prevents ugly wikitext output like
[[Foo|Foo  ]]<nowiki/>bar in VE-MW.

Bug: T54037
Change-Id: I0b734651a5b39e16142cb0edc28a594ca3d60231