Don't use a inputFilter on LinTargetInputWidget
Fixes I4df99d2c1.

Input filters run on every change making it impossible to
type space at the end of a word, even if you intend to type
something else after it.

Apply the trim only to the annotation, and make sure we don't
reset the input value from the annotation if setValue is triggered
by a change event.

Bug: T94498
Change-Id: Ie01e0f4970757f2983eb3488be805f755f688291