Update patch set 1
Patch Set 1:

(4 comments)

It's hard to tell which approach is better without knowing more about what this is going to be used for. But my guess is that this is for placing cursor holders to capture wayward cursors, in which case I feel more comfortable relying on the DOM than on the DM, out of principle.

Patch-set: 1
Label: Code-Review=0