Update patch set 2
Patch Set 2:

The problem with the DM is bidi and know where up/down takes you. CE handles that best, but skips over ce=false. We could make the ce=true all the time, but I don't like the idea of being able to place the cursor there while dragging, or accidentally hitting delete while dragging. If we're going to have ce=true somewhere where you can't really edit I want it to be for the briefest time possible.

Patch-set: 2