Allow extra properties for ve.ui.Surface
When creating a surface, allow for extra properties to be defined.
This is mostly helpful to specify whether the surface is created
within a specific dialog, especially when certain tools require
different operation when they are applied on the main surface
rather than a dialog surface. The best example of those is the
citation tools, which require a different operation when they
are activated within the reference dialog.

Bug: T94621
Change-Id: I709eeb0de47562990e709ba9d36ea0c5bd479ff6