Update patch set 2
Patch Set 2: Code-Review-1

If I understand correctly, this parses arbitrary html client-side. Which violates out security and sanitizer restrictions.

If we represent it as a <span> (like the bug suggests) then either, 1) Parsoid provides the span already and we transfer the text nodes inside of it naturally; or 2) if Parsoid provides it as a regular comment node, we just assign the DOMString nodeValue of the Comment node as the textContent or TextNode value inside the <span>.

There shouldn't be any need to decode entities manually on the client. Certainly not by using .html( .. ).text().

Patch-set: 2