ve.utils: Fix bug in batchSplice polyfill
In order to test the polyfill, we have to check the support on
each call so it can be overridden. Cache the support check so
this isn't an expensive overhead.

The actual bug was that for 'removed' we did didn't add the offset
to the second argument. Thus it was usually returning an empty array.

Change-Id: I15f9e2fa88fc53d8af40bd76228e77a53cdf7672