Update patch set 7
Patch Set 7:

It is not possible for the cursor holder to be a sibling of a ce=false node that is in a restricted context, like an <li> . Hopefully most such cases are silly/uncommon.

(In the case of tables, the whole structure should always be ce=false so this problem doesn't arise).

Patch-set: 7
Label: Verified=0