Update patch set 2
Patch Set 2:

I don't have much feedback to give. The code as I wrote it worked for me for the single case I cared about (opening the VE-MW editor). If this still works in that case (and it should, will verify and complain if it doesn't), and more, I'm all for it. :)

Patch-set: 2