Update patch set 8
Patch Set 8:

We need to prevent selections that interleave with a link (else the user can type, and kill the nails on one side only, and we can't fix up because of IME safety).

We can fixup such selections at select time, to exclude the link. For keyboard, we can modify shift+arrow to cross the entire link. For mouse, there will be a flicker but "-moz-user-select: all" can prevent this ("-webkit-user-select: all" doesn't seem to work yet).

In some cases we must apply selections to the DM without going through the DOM (e.g. Find). It looks like the DM selection should know which annotation boundaries it lies after.

Patch-set: 8
Label: Verified=0