Update patch set 1
Patch Set 1:

This should result in roughly the same colour, but it is required to make the next commit work as <hN> sets the font colour in some themes. This would also work better for a white-on-black theme or a green text theme etc.

Patch-set: 1
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0