Autolink URLs when typing.
Create ve.ui.LinkAction with an `autolinkUrl` method which autolinks
a selected URL, and a RegExp-based ve.ui.Sequence which matches
autolink-able text.

Add new `unanchoredExternalLinkUrlProtocolsRegExp` which is like the
existing `externalLinkUrlProtocolsRegExp` except it is not anchored
to the start of the string.  This allows you to execute searches for
possible links and/or add your own boundary conditions at the start
of the RegExp.  Related change I26904c85cad5de45856f1e1eea45fa14cafcbfe3
adds support for this to ve.init.mw.Platform.

Change-Id: I21976ce0cc823af29bdd2a5c163476fea2408488