Update patch set 2
Patch Set 2:

(1 comment)

At the moment I was sticking to http and friends because that makes it MW independent.  Any HTML editor will want to link those.  The MW-specific hacks I was going to put in VE-MW.  Not sure if it's important to match the MW autolink protocol list item-for-item, though?

If so I guess I can parameterize the protocol list here and let VE-MW fill it in somehow.  Ideas welcome.

Patch-set: 2