Update patch set 1
Patch Set 1:

This isn't strictly necessary for https://gerrit.wikimedia.org/r/230146 but it seems like a good idea.

As a side-effect, it uses a better property existence test so that the lookup won't crash if the type of kind is '__proto__' or 'constructor' or 'toString' or other foolishness.

Patch-set: 1
Label: Verified=0