Update patch set 19
Patch Set 19:

I've changed the logic again, so that entering/exiting the link consumes a delete/backspace press. It turns out the alternative is considerable complexity in the case of adjacent links - both for the code and for the user expectation.

Is that ok? (If so I'll update the failing tests to reflect that)

Patch-set: 19
Label: Verified=0