Update patch set 1
Patch Set 1: Code-Review+1

An interesting detail here is that (as I read it) when you cancel the progress dialog, the insertableData gets rejected as well.

In the wikitext DTH we have a failure handler which pastes the unconverted wikitext if the AJAX request fails.

I think that the way things are structured in I83ffa9e0b7164ca4a34eae86cde452d8f38dc853 we cancel the insertableData before we cancel the AJAX request, so even modulo Zalgo, we should never actually trigger the "plaintext" paste failure handler during a manual cancel.

That's probably fine: you can use "Paste Special" to bypass the automatic conversion if you need a plaintext paste.  I'm just mentioning it as a subtlety to think through.  I think the API here is fine: if you wanted the other behavior on cancel, the subclass would just not invoke the superclass abort handler (or do it only after a delay). But again, I'm mentioning this to get a second brain to think through it and double-check.

Patch-set: 1
Reviewer: C. Scott Ananian <664@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1