WIP: Fix cartouche wrapping glich with temporary block style on img
Wrapping has strange effects on the link cartouche structure:

  <img class='ve-ce-nail-pre-open' />
  <a class='ve-ce-linkAnnotation' style='box-shadow: ...'>
      <img class='ve-ce-nail-post-open' />
      ...
      <img class='ve-ce-nail-pre-close' />
  </a>
  <img class='ve-ce-nail-post-close' />

Sometimes there is a line break straight after the ve-ce-nail-post-open, which causes
an ugly "split cartouche" effect. The fix is to restyle the img.ve-ce-nail-pre-open to
block display temporarily, thereby causing a line break before the cartouche. This is
conceptually strange (as the img becomes a block element inside an inline element). It
seems to work in practice, but with quirks, e.g. cursoring down to a long wrapped link
jumps to the start of the link.

TODO decide:
1. Whether dynamic restyling is viable at all (or is there another way?)

2. Whether a block styled image inside a block is viable (it's apparently legal
in HTML+RDFa, but does splitting the block create more trouble than it solves?)

3. When exactly the styling should be updated (to turn block display on/off)

ve.ce.Surface#fixupLinks:
* fixupLinks: block style the pre-open nail of each wrapped link
* call fixupLinks after keydown / on surface observer change

ve.ce.LinkAnnotation.css:
* (debug) Style nails so they're visible and unambiguous (TODO remove)

demo/ve/pages/h1.html:
* (debug) Put link in simplest demo page (TODO remove)

Bug: T111672
Change-Id: I7ece68927dcd86a4dd7102627bf1e80685811903