Update patch set 30
Abandoned

We (Esanders and I) decided it was cleaner to allow link-bisecting selections and to tablecloth the nails away on keydown.

The cost of this is minor inconvenience (a premature commit) to some IME users, if they make a link-bisecting selection and then try to type over it. (It doesn't affect selections either wholly within a link or wholly containing a link).

Patch-set: 30
Status: abandoned