De-Houdini-fy the handling of observed DOM changes
The SurfaceObserver does far less change calculation now; it simply passes
the old and new RangeState objects to ve.ce.Surface#handleObservedChanges ,
which performs the modifications for contentChanged, selectionChanged and
branchNodeChanged in a single method.

The calculation of a ve.dm.Transaction based on a content change is moved
to a pure function ve.ce.dmChangeFromContentChange.

The cursor focus's annotation sidedness is checked when polling RangeState,
instead of at keyDown, since with IMEs there will not always be a keyDown.

Bug: T113797
Change-Id: I420c14f56187bb881c4b636a746d4383d9ba3146