GeneractedContentNode: Introduce new hash for rendering
The current model has tells you when the model has changed,
however for some model changes the rendering doesn't change.

Previously we 'fixed' this by simplifying the result of getHashObject
but this causes other problems where the model hash is used.

Instead introduce a hash specifically designed to represent the
rendered state of the node, i.e. exclude attributes which aren't visible.

Change-Id: I5d42943a06f63acab828ccad39802df573da064a