Always render generated content the first time, including errors
While using an inspector to edit an extension node, the
generated content only renders if there is no error, because
rendering the errors causes the node to keep changing size,
causing the inspector to jump around.

It is currently possible to save an extension node with errors.
If this has happened, then the error message SHOULD be displayed
when visual editor is opened, allowing the user to inspect and
correct it. Therefore a check has been added, so that generated
content is always rendered the first time, regardless of whether
there is an error. Subsequently it only rerenders if there is no
error.

Bug: T114455
Change-Id: Id16a2ff836a84cd30e15ae9d2fd894896dc46e90