MediaSizeWidget: Allow an user to really use a custom size
A custom size should allow the user to adjust the size unproportionally,
e.g. stretch an image.

This change adds a new checkbox, which is selected by default and only
activated, if the user selects a custom size, which allows the user to select,
if the opposite value (of width -> height; of height -> width) should be calculated
automatically (so that the image is proportionally scaled) or not (the user could
choose these values, he want to use).

Change-Id: I46c4c4ae1b9c4b358dda6f6bb60a012253800b73