Run renderContents before finding nodes by offset
This cleans up places which _should_ have a slug, but don't. i.e.
places which would be cleaned up anyway the next time the user did
something like cause a focus change on the surface. If this doesn't
happen, it's easy to accidentally create an almost-empty node that'll
cause errors when inspectors are used.

Bug: T110378
Change-Id: Iea6656882293e8e3cc78df87c31afe58ff59c7e8