ce.Surface afterKeyDown shouldn't assume .contenteditable is boolean
Turns out we were checking .prop('contenteditable') and assuming it'd
be a boolean, when it was really a string ('true'/'false'/'inherit').
So, switch to using .isContentEditable, which handles all that for us.

Primary impact: this now better copes with adjacent grouped nodes.
If a document contains adjacent nodes in a group, cursor navigation
would only sometimes treat them as a group. Moving to after the group
and pressing left, in particular, would easily get the cursor in the
middle of the group. The cursor is _technically_ not in a non-editable
location at this time, but in practice it should be treated as being
so.

Bug: T114545
Change-Id: I4452aa05753fdacd29b861c058ec62f9ca918a8f