ce.Surface afterKeyDown to better cope with adjacent grouped nodes
If a document contains adjacent nodes in a group, cursor navigation
would only sometimes treat them as a group. Moving to after the group
and pressing left, in particular, would easily get the cursor in the
middle of the group. The cursor is _technically_ not in a non-editable
location at this time, but in practice it should be treated as being
so.

Therefore, adjust getSurroundingFocusableNode to notice when the node
it's checking could be focused in this manner.

Bug: T114545
Change-Id: I4452aa05753fdacd29b861c058ec62f9ca918a8f